Fix some demo bugs (#30)
* numeric equality bugfix, file path fix, support ~ in paths

* fix

* allow no condition

* change extension to fldr

* adjustments

---------

+++ src/ast/condition/comparison/numeric/NumericComparisonType.java
@@ -7,7 +7,7 @@ public enum NumericComparisonType {
     GREATER_THAN_EQUAL((a, b) -> a >= b),
     LESS_THAN((a, b) -> a < b),
     LESS_THAN_EQUAL((a, b) -> a <= b),
-    EQUAL_TO((a,b) -> a == b);
+    EQUAL_TO((a,b) -> (long) a == (long) b);
 
     private BiFunction<Long, Long, Boolean> comparisonFunction;
